Relocate fixtures for testing and FakePrismLookup class to a separate conftest file
